Extract frontend-neutral turn lifecycle into shared runtime
